+++ tests/unit/CRUD/base/base-reading-simple-predicates-test.js
@@ -2,9 +2,10 @@ import Ember from 'ember';
 import QueryBuilder from 'ember-flexberry-data/query/builder';
 import FilterOperator from 'ember-flexberry-data/query/filter-operator';
 import { SimplePredicate } from 'ember-flexberry-data/query/predicate';
+import { ConstParam, AttributeParam } from 'ember-flexberry-data/query/parameter';
 
 export default function readingPredicatesSimplePredicatesOperators(store, assert) {
-  assert.expect(24);
+  assert.expect(72);
   let done = assert.async();
 
   Ember.run(() => {
@@ -31,6 +32,42 @@ export default function readingPredicatesSimplePredicatesOperators(store, assert
       );
     })
 
+    // Eq among AttributeParam and ConstParam.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('karma'), '==', new ConstParam(5));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('karma'), FilterOperator.Eq, new ConstParam(5)));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Eq among AttributeParam and ConstParam with operator | Data', 'Eq among AttributeParam and ConstParam with operator | Length'],
+          ['Eq among AttributeParam and ConstParam with simple predicate | Data', 'Eq among AttributeParam and ConstParam with simple predicate | Length']
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('karma') === 5), messages[0]);
+          assert.equal(data.get('length'), 2, messages[1]);
+        }
+      );
+    })
+
+    // Eq among AttributeParams.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '==', new AttributeParam('eMail'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Eq, new AttributeParam('eMail')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Eq among AttributeParams with operator | Data', 'Eq among AttributeParams with operator | Length'],
+          ['Eq among AttributeParams with simple predicate | Data', 'Eq among AttributeParams with simple predicate | Length']
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') === item.get('eMail')), messages[0]);
+          assert.equal(data.get('length'), 1, messages[1]);
+        }
+      );
+    })
+
     // Neq.
     .then(() => {
       builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
@@ -44,7 +81,43 @@ export default function readingPredicatesSimplePredicatesOperators(store, assert
         ],
         (data, messages) => {
           assert.ok(data.every(item => item.get('karma') !== 5), messages[0]);
-          assert.equal(data.get('length'), 2, messages[1]);
+          assert.equal(data.get('length'), 3, messages[1]);
+        }
+      );
+    })
+
+    // Neq among AttributeParam and ConstParam.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('karma'), '!=', new ConstParam(5));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('karma'), FilterOperator.Neq, new ConstParam(5)));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Neq among AttributeParam and ConstParam with operator | Data', 'Neq among AttributeParam and ConstParam with operator | Length'],
+          ['Neq among AttributeParam and ConstParam with simple predicate | Data', 'Neq among AttributeParam and ConstParam with simple predicate | Length']
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('karma') !== 5), messages[0]);
+          assert.equal(data.get('length'), 3, messages[1]);
+        }
+      );
+    })
+
+    // Neq among AttributeParams.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '!=', new AttributeParam('eMail'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Neq, new AttributeParam('eMail')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Neq with operator | Data', 'Neq with operator | Length'],
+          ['Neq with simple predicate | Data', 'Neq with simple predicate | Length']
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') !== item.get('eMail')), messages[0]);
+          assert.equal(data.get('length'), 4, messages[1]);
         }
       );
     })
@@ -62,7 +135,43 @@ export default function readingPredicatesSimplePredicatesOperators(store, assert
         ],
         (data, messages) => {
           assert.ok(data.every(item => item.get('karma') > 4), messages[0]);
-          assert.equal(data.get('length'), 3, messages[1]);
+          assert.equal(data.get('length'), 4, messages[1]);
+        }
+      );
+    })
+
+    // Ge among AttributeParam and ConstParam.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '>', new ConstParam('Andrey'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Ge, new ConstParam('Andrey')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Ge among AttributeParam and ConstParam with operator | Data', 'Ge among AttributeParam and ConstParam with operator | Length'],
+          ['Ge among AttributeParam and ConstParam with simple predicate | Data', 'Ge among AttributeParam and ConstParam with simple predicate | Length'],
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') > 'Andrey'), messages[0]);
+          assert.equal(data.get('length'), 4, messages[1]);
+        }
+      );
+    })
+
+    // Ge among AttributeParams.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '>', new AttributeParam('eMail'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Ge, new AttributeParam('eMail')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Ge among AttributeParams with operator | Data', 'Ge among AttributeParams with operator | Length'],
+          ['Ge among AttributeParams with simple predicate | Data', 'Ge among AttributeParams with simple predicate | Length'],
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') > item.get('eMail')), messages[0]);
+          assert.equal(data.get('length'), 4, messages[1]);
         }
       );
     })
@@ -80,7 +189,43 @@ export default function readingPredicatesSimplePredicatesOperators(store, assert
         ],
         (data, messages) => {
           assert.ok(data.every(item => item.get('karma') >= 5), messages[0]);
-          assert.equal(data.get('length'), 3, messages[1]);
+          assert.equal(data.get('length'), 4, messages[1]);
+        }
+      );
+    })
+
+    // Geq among AttributeParam and ConstParam.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '>=', new ConstParam('Kolya'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Geq, new ConstParam('Kolya')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Geq among AttributeParam and ConstParam with operator | Data', 'Geq among AttributeParam and ConstParam with operator | Length'],
+          ['Geq among AttributeParam and ConstParam with simple predicate | Data', 'Geq among AttributeParam and ConstParam with simple predicate | Length'],
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') >= 'Kolya'), messages[0]);
+          assert.equal(data.get('length'), 4, messages[1]);
+        }
+      );
+    })
+
+    // Geq among AttributeParams.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '>=', new AttributeParam('eMail'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Geq, new AttributeParam('eMail')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Geq among AttributeParams with operator | Data', 'Geq among AttributeParams with operator | Length'],
+          ['Geq among AttributeParams with simple predicate | Data', 'Geq among AttributeParams with simple predicate | Length'],
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') >= item.get('eMail')), messages[0]);
+          assert.equal(data.get('length'), 5, messages[1]);
         }
       );
     })
@@ -103,6 +248,42 @@ export default function readingPredicatesSimplePredicatesOperators(store, assert
       );
     })
 
+    // Le among AttributeParam and ConstParam.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '<', new ConstParam('Kolya'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Le, new ConstParam('Kolya')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Le among AttributeParam and ConstParam with operator | Data', 'Le among AttributeParam and ConstParam with operator | Length'],
+          ['Le among AttributeParam and ConstParam with simple predicate | Data', 'Le among AttributeParam and ConstParam with simple predicate | Length'],
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') < 'Kolya'), messages[0]);
+          assert.equal(data.get('length'), 1, messages[1]);
+        }
+      );
+    })
+
+    // Le among AttributeParams.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '<', new AttributeParam('eMail'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Le, new AttributeParam('eMail')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Le among AttributeParams with operator | Data', 'Le among AttributeParams with operator | Length'],
+          ['Le among AttributeParams with simple predicate | Data', 'Le among AttributeParams with simple predicate | Length'],
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') < item.get('eMail')), messages[0]);
+          assert.equal(data.get('length'), 0, messages[1]);
+        }
+      );
+    })
+
     // Leq.
     .then(() => {
       builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
@@ -120,6 +301,43 @@ export default function readingPredicatesSimplePredicatesOperators(store, assert
         }
       );
     })
+
+    // Leq among AttributeParam and ConstParam.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '<=', new ConstParam('Kolya'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Leq, new ConstParam('Kolya')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Leq among AttributeParam and ConstParam with operator | Data', 'Leq among AttributeParam and ConstParam with operator | Length'],
+          ['Leq among AttributeParam and ConstParam with simple predicate | Data', 'Leq among AttributeParam and ConstParam with simple predicate | Length'],
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') <= 'Kolya'), messages[0]);
+          assert.equal(data.get('length'), 2, messages[1]);
+        }
+      );
+    })
+
+    // Leq among AttributeParams.
+    .then(() => {
+      builderStrOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new AttributeParam('name'), '<=', new AttributeParam('eMail'));
+      builderConstOp = new QueryBuilder(store, 'ember-flexberry-dummy-application-user')
+        .where(new SimplePredicate(new AttributeParam('name'), FilterOperator.Leq, new AttributeParam('eMail')));
+
+      return runTest(store, [builderStrOp, builderConstOp], [
+          ['Leq among AttributeParams with operator | Data', 'Leq among AttributeParams with operator | Length'],
+          ['Leq among AttributeParams with simple predicate | Data', 'Leq among AttributeParams with simple predicate | Length'],
+        ],
+        (data, messages) => {
+          assert.ok(data.every(item => item.get('name') <= item.get('eMail')), messages[0]);
+          assert.equal(data.get('length'), 1, messages[1]);
+        }
+      );
+    })
+
     .catch((e) => {
       console.log(e, e.message);
       throw e;
@@ -152,6 +370,12 @@ function initTestData(store) {
       name: 'Andrey',
       eMail: '4@mail.ru',
       karma: 6
+    }).save(),
+
+    store.createRecord('ember-flexberry-dummy-application-user', {
+      name: 'SameData',
+      eMail: 'SameData',
+      karma: 7
     }).save()
   ]);
 }