+++ tests/unit/CRUD/odata/odata-reading-simple-predicates-with-first-level-masters-test.js
@@ -1,0 +1,6 @@
+import executeTest from './execute-odata-test';
+import readingPredicatesSimplePredicatesWithMastersOperators from '../base/base-reading-simple-predicates-with-first-level-masters-test';
+
+executeTest('reading with first-level-masters | predicates | simple predicates | operators', (store, assert) => {
+  readingPredicatesSimplePredicatesWithMastersOperators(store, assert);
+});